gc stats: move the code to its own file gc_stats.{c,h}
